separando arquivos, criaçao do pipfile e do setup.py
