+++ 05 Análise de Texto/PyCharm/entregas/Pipfile
@@ -1,0 +1,17 @@
+[[source]]
+url = "https://pypi.org/simple"
+verify_ssl = true
+name = "pypi"
+
+[packages]
+spacy = "*"
+gensim = "==3.8.3"
+nltk = "*"
+pysub-parser = "*"
+matplotlib = "*"
+wordcloud = "*"
+
+[dev-packages]
+
+[requires]
+python_version = "3.9"
