+++ 05 Análise de Texto/PyCharm/entregas/setup.py
@@ -1,0 +1,8 @@
+import spacy
+import nltk
+
+nltk.download('averaged_perceptron_tagger')
+nltk.download('stopwords')
+
+spacy.cli.download('en_core_web_sm')
+spacy.cli.download('en_core_web_trf')
